+++ app/helpers/chaltron/logs_helper.rb
@@ -6,10 +6,14 @@ module Chaltron::LogsHelper
 
   def bootstrap_severity(severity)
     case severity.to_s
-    when 'info'  then 'primary'
-    when 'error' then 'danger'
-    when 'debug' then 'info'
-    else 'default'
+    when 'emerg', 'alert', 'crit', 'err'
+      'danger'
+    when 'warning'
+      'warning'
+    when 'debug'
+      'info'
+    else
+      'primary'
     end
   end
 end