+++ spec/models/log_spec.rb
@@ -10,8 +10,15 @@ RSpec.describe Log, type: :model do
       subject { log }
       let(:log) { build(:log, severity: severity) }
       context 'good' do
-        let(:severity) { 'info' }
+        let(:severity) { 'notice' }
         it { is_expected.to be_valid }
+      end
+      context 'syslog' do
+        let(:severity) { 'error' }
+        it do
+          is_expected.to be_valid
+          expect(log.severity).to eq('err')
+        end
       end
       context 'bad' do
         let(:severity) { 'bad' }